added controllers model and routes
