+++ src/books/model.js
@@ -1,0 +1,19 @@
+const mongoose = require("mongoose");
+
+const bookSchema = new mongoose.Schema({
+  title: {
+    type: String,
+    required: true,
+    unique: true,
+  },
+  author: {
+    type: String,
+  },
+  genre: {
+    type: String,
+  },
+});
+
+const Book = mongoose.model("Book", bookSchema);
+
+module.exports = Book;
